More kernel tweaks (#6428)
* kernel: drop unused return value of UseTmpPerm

* kernel: use BOOL in more places

* kernel: init buffer with memset

* kernel: use size_t in some places

* kernel: explain certain uses of 'extern inline'

* kernel: update some comments

+++ src/records.c
@@ -64,7 +64,7 @@ static void NewRNamCallback(SymbolTable * symtab, UInt id, Obj name)
 */
 UInt RNamName(const Char * name)
 {
-    UInt len = strlen(name);
+    size_t len = strlen(name);
     if (len > 1023) {
         // Note: We can't pass 'name' here, as it might get moved by garbage collection
         ErrorQuit("Record names must consist of at most 1023 characters", 0, 0);
